ManagedEsent v1.8. Added Windows 8 and Metro support. Added a new Windows8Api class and appropriate namespace/grbits.
The support for Metro is present, but it is not pretty:
-It is not possible to have a single DLL support XP and Metro, so there are now two binaries being built.
-Many ifdef's all over the place. Very messy.
-The tests do not clean up after themselves.
-I could not get the integrated tests to compile with VS, so I ended up duplicating the code in VisualStudioUnitTest.cs by using Reflector.
-No attempt was made to port PersistentDictionary.

The test application I used to verify the changes is not checked in. Here are the useful functions if you'd like to run them:
-TestRunner.MainClass.TestRunnerMain("InteropApiTests", "-test", testname, "-donotcatch");
-TestRunner.MainClass.TestRunnerMain("InteropApiTests", "-priority", priority.ToString(), "-donotcatch");
-TestRunner.MainClass.TestRunnerMain("InteropApiTests", "-suite", suiteName, "-donotcatch");
The output goes to the Debug output, since there is no Console in Metro.

Some (50) of the tests for non-Metro will fail on Win8 Beta, because they rely on esent.dll changes that will be in RC.