Merged PR 497050: Remove very expensive assert in ESE interop  layer
The ESE interop layer has a TraceFunctionCall() method that is intended to note every ESE interop method that is called.  This method requires passing the exact name of the invoking method and it verifies this via a very expensive assert that takes a stack dump.  This is prohibitively expensive and is totally unnecessary.

This change uses CallerMemberName to reliably implement the exact same functionality more cheaply and more reliably.
[Substrate:8d3f087058d48]