Merged PR 608902: Add type name string to  PersistentDictionary Globals table for validating primitive types
PersistentDictionary used to have code that would serialize the type and save it in the Globals table to validate on re-open that the same types were used. The serialization code (along with support for serializable structs in PD) was removed from Exchange. This change adds new columns in the Globals table that store the type name so validation can still happen for primitive types.

The upgrade is done at startup by checking for the existence of the new columns and if not found, adding them and saving the type names.
[Substrate:cd55acbe5eb55]