Remove DynamicMethod from PersistentDictionary to make it  compatible with .NET
Merged PR 626659: Remove DynamicMethod from PersistentDictionary to make it compatible with .NET Standard 2.0.

There isn't any real functional change here. Previously, the code created a DynamicMethod that called the EsentInterop API and got back a nullable value, then it called the .Value() accessor on that object. This code simply creates the equivalent functions. It's slightly less elegant, but with this PersistentDictionary can be part of .NET Standard 2.0.
[Substrate:528708916109f]