Merged PR 735870: ESE BlockCache File Handle Table
This change introduces the File Handle Table.  This table is used to allow the engine to continue to manipulate files normally while also allowing the Block Cache to open files in the same process for write back.  It does this by creating virtual file handles (which are exposed as IFileAPI*) and maintaining the proper file access/sharing semantics of the underlying file system except it allows an extra open of a file as RW Exclusive for write back.  This will also allow us (eventually) to silently upgrade an open of a RO Shared file to RW Exclusive to support write back if the file is cached.

Note that we do not do automatic retries until the timeout to create/open a file if there is a conflict in the file handle table.  We will immediately fail.

To support the file access checks, we have factored out the methods used in OSFS to translate IFileAPI::FileModeFlags into the native OS flags.  The implementation will be shared with the BlockCache.

This change also adds the OS interactions required to manipulate file ids and to map them to file paths.  This is basically the only new OS interaction required for the entire feature.  We will bump ESE's min OS version to 0x0603 as a part of this change.

We also have moved most of the BlockCache tests into a new test DLL called Internal.Ese.QTests.DLL.  This is where we will put all ESE QTests in the future so that we can preserve legacy functional testing in the existing test DLLs.  These tests should otherwise execute normally in all ways (visual studio, qtest, accept, perseus) and tests are added using the usual SyncTool mechanism.  The functional BlockCache tests will still live in WSTF.Tests.

ESE Block Cache v1 Design: https://microsoft-my.sharepoint-df.com/:w:/p/andygo_ntdev/EdAsUsjibz5Pt9aoBifkR6gBhXPFCgLVdsdz10qk15xydQ?e=W3fccp

Related work items: #1445635
[Substrate:4bcf3c21146cc]