Merged PR 763980: ESE Block Cache Configuration
This change provides the production configuration for the ESE Block Cache.  In essence, we are creating a mapping between files we want to cache and the files that back this cache via Variant Configuration.  This configuration is consumed by ESE via some new interfaces.  This configuration is provided to ESE via C# interop with Variant Configuration.  The plumbing for the interop is provided by BlockCache.Interop and the configuration lives in ExchangeDatabaseManager and is loaded along with the other ESE configuration per instance when we are invoked by Store.Worker, HA, etc.

Each type of cached file and caching file has further configuration to control its behavior as described in the design.  This configuration is simple at the moment but may have more settings later.

There is also a master switch that can turn off 100% of the block cache code even in the configuration paths.

Some example configuration has been added that expresses the caching we wish to someday have for the various types of files used by the database copies on the BE.  This caching is configured to be off.  Note that we could also configure any other ESE database for caching using this same technique.  Also note that we had to add code to "abstract" the database name, base name, and log generation from the paths so that we could get stable paths to configure for the database copy files.

Other notable changes:
-  to enable the use of BlockCache.Interop by ExchangeDatabaseManager we had to change how some of ESE is built with respect to build phases.  Sync.lib and oslite.lib are now linked in pass 1.  ExchangeDatabaseManager is no longer linked in phase 1 (and this is no longer necessary for build).
-  ExchangeDatabaseManager holds a cache of native resources that must be freed on process exit to avoid memory leak asserts for tests.  It accomplishes this via a process exit handler.  The TestFlight code was using a Win32 interop call to exit the process which bypasses a clean CLR process exit, defeating this cleanup.  This technique is unnecessary anyway as the error code should simply be returned from Main.

ESE Block Cache V1 Design:  https://microsoft.sharepoint-df.com/:w:/t/EXOmi/ERis-UtREBFHrRRD7SCXVW0BF3OxVDpPg3V83zicTbgD1g
[Substrate:78ed2246c042f]