Merged PR 797165: Migrate ESE tests to MSTest2
We have a bunch of different versions of Microsoft.VisualStudio.QualityTools.UnitTestFramework in Substrate today.  Trying to move everything forward to the latest MSTest2 version (Microsoft.VisualStudio.QualityTools.UnitTestFramework in the MSTest.TestFramework package).  All of Transport and many other projects have been migrated already when they took a dependency on it as a part of another change.  Migrating ESE tests separately since they have special logic that runs the tests in a wrapper and external TestFlight.exe process.
[Substrate:c9592d999e334]