Merged PR 869107: JetGetSystemParameter interop is  inefficient when called at high concurrency
We hit an issue in prod where many concurrent calls to SystemParameter.KeyMost made by store were consuming huge amounts of CPU spinning on a lock in string.IsInterned sadly while trying to intern string.Empty.

It is reasonable to expect people to make this kind of info call frequently.

This change:
-  optimizes the memory usage in JetGetSystemParameter interop by using a stackalloc buffer rather than a StringBuilder which will avoid allocations when fetching a parameter that doesn't return a string
-  avoids the bad performance of string.IsInterned for null/string.IsEmpty
[Substrate:3681d0f775181]