Merged PR 913055: Fix managed JET_INDEXID field order
In looking into using the JET table objid for per-table usage tracking, I noticed that the managed and native definitions for the JET_INDEXID structure didn't match. The native structure defines the two uints (cbStruct and objidFPD) before the IntPtr (with a comment that this preserves the 8 byte alignment for the pointer on x64, which makes me wonder if the managed struct just matches an older version of the native struct that predates that comment).

At any rate, this change fixes the managed definition and renames the fields in the managed struct to make it easier to identify the objid.

(Since the the table ID matches the index ID of the primary index, @<Jonathan Liem> recommended using this API to get the table ID.)

I've also added a test to ensure the objid retrieved from Store Query matches the one retrieved from this API. (Without the fix, the objid appears to be half of the managed IntPtr.)
[Substrate:18cfcf798e13b]