Merged PR 2427852: Multi-targeting EsentCollections and isam  to .NetCore
**Motivation**
We plan to leverage PersistentDictionary in our .NetCore project. Current `microsoft.isam.esent.collections` is only available in .Net Framework. It will be port to .NetCore with multi-targeting in this PR.

**Changes**
1.`microsoft.isam.esent.collections` is multi-targeting to support .NetCore 3.1 & 6.0.
2.Since `microsoft.isam.esent.collections` refers `Microsoft.Database.Isam` which is also not available in .NetCore, `Microsoft.Database.Isam` is also port to .NetCore 3.1 & 6.0 using multi-targeting.
3.After porting above two projects to .NetCore, we have add `SetTargetFramework` tag to other projects that refers the above two projects.

**Tests**
We will rely on focus job and existing unit tests to verify the migration.
Currently `microsoft.isam.esent.collections` is only used in SharedCache and cafe team deprecated `PersistentDatabase` that used `microsoft.isam.esent.collections`, so this should be a safe change.

Related work items: #3288617
[Substrate:c15c01d261fb635067de3ca7f4ac83e7c3d8fac3]